Revert "Split utils/validation.py kitchen-sink into focused modules"
