+++ utils/validation.py
@@ -1,32 +1,15 @@
-"""Validation orchestration and field type registration.
-
-Column validators live in :mod:`utils.validators`.
-Value normalizers live in :mod:`utils.normalizers`.
-This module wires them together into the field type registry and exposes the
-:func:`validation_sorter` dispatcher used by views.
-"""
-
+import re
+import csv
 import logging
 from db.schema import get_field_schema
 from utils.field_registry import register_type, get_field_type
-from utils.validators import (
-    validate_text_column,
-    validate_textarea_column,
-    validate_number_column,
-    validate_boolean_column,
-    validate_select_column,
-    validate_multi_select_column,
-)
-from utils.normalizers import normalize_boolean, normalize_number, normalize_multi
 
 logger = logging.getLogger(__name__)
-
 
 def get_options(table: str, field: str) -> list[str]:
     """Return options for the given field from the field schema."""
     schema = get_field_schema()
     return schema.get(table, {}).get(field, {}).get("options", [])
-
 
 def validation_sorter(table, field, header, fieldType, values):
     logger.debug(
@@ -58,7 +41,174 @@ def validation_sorter(table, field, header, fieldType, values):
         result["cssClass"] = " ".join(classes)
 
     return result
+        
 
+
+def validate_text_column(values):
+    valid = invalid = blank = warning = 0
+    details = {"blank": [],"invalid":[],"warning":[],"valid":[]}
+    for idx, v in enumerate(values, start=1):
+        if not v or v.strip() == "":
+            blank += 1
+            details["blank"].append(idx)
+            continue
+        if len(v) > 1000:
+            invalid += 1
+            details["invalid"].append({"row": idx, "reason": "length exceeds 1000 characters"})
+        else:
+            valid += 1
+            details["valid"].append(idx)
+        # warning check based on common identifiers 
+        if any([
+            re.search(r"\{.*?\}", v),                 # JSON
+            re.search(r"<[a-z][\s\S]*?>", v),         # HTML tags
+            re.search(r"#+\s", v)                     # markdown headers
+        ]):
+            warning += 1
+            details["warning"].append({"row": idx, "reason": "contains JSON,HTML, or MD"})
+    return {
+        "valid": valid,
+        "invalid": invalid,
+        "blank": blank,
+        "warning": warning,
+        "details": details
+    }
+def validate_textarea_column(values):
+    max_size = csv.field_size_limit()  # default ~131072 bytes
+    details = {"blank": [],"invalid":[],"warning":[],"valid":[]}
+    valid = invalid = blank = warning = 0
+    for idx, v in enumerate(values, start=1):
+        if not v or v.strip() == "":
+            blank += 1
+            details["blank"].append(idx)
+            continue
+        # Invalid if over CSV field limit
+        if len(v) > max_size:
+            invalid += 1
+            details["invalid"].append({"row": idx, "reason": f"length exceeds {max_size} characters"})
+        else:
+            valid += 1
+            details["valid"].append(idx)
+        # Warning: JSON-like content
+        if re.search(r"\{.*?\}", v):
+            warning += 1
+            details["warning"].append({"row": idx, "reason": "Possibly contains JSON"})
+    return {
+        "valid": valid,
+        "invalid": invalid,
+        "blank": blank,
+        "warning": warning,
+        "details": details
+    }
+def validate_number_column(values, integer_only=False):
+    valid = invalid = blank = 0
+    details = {"blank": [],"invalid":[],"warning":[],"valid":[]}
+    for idx, v in enumerate(values, start=1):
+        if not v or v.strip() == "":
+            blank += 1
+            details["blank"].append(idx)
+            continue
+        s = str(v).strip()
+        # Try to convert to number
+        try:
+            int(s) if integer_only else float(s)
+        except ValueError:
+            logger.debug("Invalid numeric value", exc_info=True)
+            invalid += 1
+            details["invalid"].append({"row": idx, "reason": "not a number", })
+            continue
+        else:
+            valid += 1
+            details["valid"].append(idx)
+    return {
+        "valid": valid,
+        "invalid": invalid,
+        "blank": blank,
+        "details": details
+    }
+def validate_boolean_column(values):
+    valid = invalid = blank = 0
+    details = {"blank": [],"invalid":[],"warning":[],"valid":[]}
+    for idx, v in enumerate(values, start=1):
+        if not v or v.strip() == "":
+            blank += 1
+            details["blank"].append(idx)
+            continue
+        s = str(v).strip().lower()
+        if s in ("true", "false", "1", "0", "yes", "no"):
+            valid += 1
+            details["valid"].append(idx)
+        else:
+            invalid += 1
+            details["invalid"].append({"row": idx, "reason": "invalid boolean value","value":v })
+    return {"valid": valid, "invalid": invalid, "blank": blank, "details":details}
+def validate_select_column(values: list[str], options: list[str]) -> dict:
+    valid = invalid = blank = 0
+    normalized_options = {opt.lower() for opt in options}
+    details = {"blank": [],"invalid":[],"warning":[],"valid":[]}
+    for idx, v in enumerate(values, start=1):
+        if not v or v.strip() == "":
+            blank += 1
+            details["blank"].append(idx)
+        elif str(v).strip().lower() in normalized_options:
+            valid += 1
+            details["valid"].append(idx)
+        else:
+            invalid += 1
+            details["invalid"].append({"row": idx, "reason": "does not match available options","value":v })
+    return {"valid": valid, "invalid": invalid, "blank": blank, "details":details}
+def validate_multi_select_column(values: list[str], options: list[str]) -> dict:
+    valid = invalid = blank = warning = 0
+    details = {"valid": [],"invalid": [],"blank": [],"warning": []}
+    # Prepare lookup
+    normalized_options = {opt.lower() for opt in options}
+    for idx, raw in enumerate(values, start=1):
+        cell = str(raw).strip()
+        # Blank cell
+        if not cell:
+            blank += 1
+            details["blank"].append(idx)
+            continue
+        # Split into individual tags
+        tags = [t.strip() for t in cell.split(',')]
+        # Identify any tags not in the allowed list
+        invalid_tags = [t for t in tags if t.lower() not in normalized_options]
+        if not invalid_tags:
+            valid += 1
+            details["valid"].append(idx)
+        else:
+            invalid += 1
+            details["invalid"].append({
+                "row": idx,
+                "value": raw,
+                "reason": f"invalid tags: {invalid_tags}"
+            })
+
+    return {
+        "valid": valid,
+        "invalid": invalid,
+        "blank": blank,
+        "warning": warning,
+        "details": details
+    }
+
+
+def normalize_boolean(value):
+    return "1" if str(value).lower() in {"1", "true", "on", "yes"} else "0"
+
+
+def normalize_number(value):
+    try:
+        return str(float(value))
+    except (TypeError, ValueError):
+        logger.warning("Failed to normalize number", exc_info=True)
+        return "0"
+
+
+def normalize_multi(value):
+    if isinstance(value, (list, tuple, set)):
+        return ", ".join(str(v) for v in value)
+    return "" if value is None else str(value)
 
 # Register built-in field types with the registry
 register_type(
