--- utils/validators.py
@@ -1,162 +1,0 @@
-"""Column-level validators for CSV import and data validation.
-
-Each function accepts a list of raw string values and returns a dict with
-counts (blank, valid, invalid, warning) and a details breakdown.
-"""
-
-import re
-import csv
-import logging
-
-logger = logging.getLogger(__name__)
-
-
-def validate_text_column(values):
-    valid = invalid = blank = warning = 0
-    details = {"blank": [], "invalid": [], "warning": [], "valid": []}
-    for idx, v in enumerate(values, start=1):
-        if not v or v.strip() == "":
-            blank += 1
-            details["blank"].append(idx)
-            continue
-        if len(v) > 1000:
-            invalid += 1
-            details["invalid"].append({"row": idx, "reason": "length exceeds 1000 characters"})
-        else:
-            valid += 1
-            details["valid"].append(idx)
-        # warning check based on common identifiers
-        if any([
-            re.search(r"\{.*?\}", v),                # JSON
-            re.search(r"<[a-z][\s\S]*?>", v),        # HTML tags
-            re.search(r"#+\s", v)                    # markdown headers
-        ]):
-            warning += 1
-            details["warning"].append({"row": idx, "reason": "contains JSON, HTML, or MD"})
-    return {
-        "valid": valid,
-        "invalid": invalid,
-        "blank": blank,
-        "warning": warning,
-        "details": details,
-    }
-
-
-def validate_textarea_column(values):
-    max_size = csv.field_size_limit()  # default ~131072 bytes
-    details = {"blank": [], "invalid": [], "warning": [], "valid": []}
-    valid = invalid = blank = warning = 0
-    for idx, v in enumerate(values, start=1):
-        if not v or v.strip() == "":
-            blank += 1
-            details["blank"].append(idx)
-            continue
-        if len(v) > max_size:
-            invalid += 1
-            details["invalid"].append({"row": idx, "reason": f"length exceeds {max_size} characters"})
-        else:
-            valid += 1
-            details["valid"].append(idx)
-        if re.search(r"\{.*?\}", v):
-            warning += 1
-            details["warning"].append({"row": idx, "reason": "Possibly contains JSON"})
-    return {
-        "valid": valid,
-        "invalid": invalid,
-        "blank": blank,
-        "warning": warning,
-        "details": details,
-    }
-
-
-def validate_number_column(values, integer_only=False):
-    valid = invalid = blank = 0
-    details = {"blank": [], "invalid": [], "warning": [], "valid": []}
-    for idx, v in enumerate(values, start=1):
-        if not v or v.strip() == "":
-            blank += 1
-            details["blank"].append(idx)
-            continue
-        s = str(v).strip()
-        try:
-            int(s) if integer_only else float(s)
-        except ValueError:
-            logger.debug("Invalid numeric value", exc_info=True)
-            invalid += 1
-            details["invalid"].append({"row": idx, "reason": "not a number"})
-            continue
-        else:
-            valid += 1
-            details["valid"].append(idx)
-    return {
-        "valid": valid,
-        "invalid": invalid,
-        "blank": blank,
-        "details": details,
-    }
-
-
-def validate_boolean_column(values):
-    valid = invalid = blank = 0
-    details = {"blank": [], "invalid": [], "warning": [], "valid": []}
-    for idx, v in enumerate(values, start=1):
-        if not v or v.strip() == "":
-            blank += 1
-            details["blank"].append(idx)
-            continue
-        s = str(v).strip().lower()
-        if s in ("true", "false", "1", "0", "yes", "no"):
-            valid += 1
-            details["valid"].append(idx)
-        else:
-            invalid += 1
-            details["invalid"].append({"row": idx, "reason": "invalid boolean value", "value": v})
-    return {"valid": valid, "invalid": invalid, "blank": blank, "details": details}
-
-
-def validate_select_column(values: list[str], options: list[str]) -> dict:
-    valid = invalid = blank = 0
-    normalized_options = {opt.lower() for opt in options}
-    details = {"blank": [], "invalid": [], "warning": [], "valid": []}
-    for idx, v in enumerate(values, start=1):
-        if not v or v.strip() == "":
-            blank += 1
-            details["blank"].append(idx)
-        elif str(v).strip().lower() in normalized_options:
-            valid += 1
-            details["valid"].append(idx)
-        else:
-            invalid += 1
-            details["invalid"].append({"row": idx, "reason": "does not match available options", "value": v})
-    return {"valid": valid, "invalid": invalid, "blank": blank, "details": details}
-
-
-def validate_multi_select_column(values: list[str], options: list[str]) -> dict:
-    valid = invalid = blank = warning = 0
-    details = {"valid": [], "invalid": [], "blank": [], "warning": []}
-    normalized_options = {opt.lower() for opt in options}
-    for idx, raw in enumerate(values, start=1):
-        cell = str(raw).strip()
-        if not cell:
-            blank += 1
-            details["blank"].append(idx)
-            continue
-        tags = [t.strip() for t in cell.split(',')]
-        invalid_tags = [t for t in tags if t.lower() not in normalized_options]
-        if not invalid_tags:
-            valid += 1
-            details["valid"].append(idx)
-        else:
-            invalid += 1
-            details["invalid"].append({
-                "row": idx,
-                "value": raw,
-                "reason": f"invalid tags: {invalid_tags}",
-            })
-    return {
-        "valid": valid,
-        "invalid": invalid,
-        "blank": blank,
-        "warning": warning,
-        "details": details,
-    }
